Check SSL file permissions when creating an ssl_context
