+++ chia/ssl/create_ssl.py
@@ -4,15 +4,13 @@ from pathlib import Path
 from typing import Any, List, Tuple
 
 import pkg_resources
+from chia.util.ssl import DEFAULT_PERMISSIONS_CERT_FILE, DEFAULT_PERMISSIONS_KEY_FILE
 from cryptography import x509
 from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import hashes, serialization
 from cryptography.hazmat.primitives.asymmetric import rsa
 from cryptography.hazmat.primitives.serialization import load_pem_private_key
 from cryptography.x509.oid import NameOID
-
-DEFAULT_PERMISSIONS_CERT_FILE = 0o644
-DEFAULT_PERMISSIONS_KEY_FILE = 0o600
 
 
 def get_chia_ca_crt_key() -> Tuple[Any, Any]:
